update the graphql

+++ models/user.go
@@ -7,7 +7,7 @@ import (
 )
 
 type User struct {
-	Id        int64      `gorm:"primary_key" db:"id" json:"id"`
+	Id        int        `gorm:"primary_key" db:"id" json:"id"`
 	CreatedAt time.Time  `db:"created_at" json:"createdAt"`
 	UpdatedAt time.Time  `db:"updated_at" json:"updatedAt"`
 	DeletedAt *time.Time `db:"deleted_at" json:"deletedAt"`
